+++ apps/meteor/app/apps/server/converters/videoConferences.ts
@@ -1,6 +1,6 @@
 import type { VideoConference } from '@rocket.chat/apps-engine/definition/videoConferences';
-import type { IVideoConference } from '@rocket.chat/core-typings';
 import { VideoConf } from '@rocket.chat/core-services';
+import type { IVideoConference } from '@rocket.chat/core-typings';
 
 export class AppVideoConferencesConverter {
 	async convertById(callId: string): Promise<VideoConference | undefined> {