+++ apps/meteor/app/ui-message/client/ActionManager.js
@@ -1,18 +1,17 @@
 import { UIKitIncomingInteractionType } from '@rocket.chat/apps-engine/definition/uikit';
-import { Meteor } from 'meteor/meteor';
-import { Random } from '@rocket.chat/random';
-import { Emitter } from '@rocket.chat/emitter';
 import { UIKitInteractionTypes } from '@rocket.chat/core-typings';
+import { Emitter } from '@rocket.chat/emitter';
+import { Random } from '@rocket.chat/random';
+import { lazy } from 'react';
 
-import Notifications from '../../notifications/client/lib/Notifications';
-import { CachedCollectionManager } from '../../ui-cached-collection/client';
-import { t } from '../../utils/client';
 import * as banners from '../../../client/lib/banners';
-import { dispatchToastMessage } from '../../../client/lib/toast';
 import { imperativeModal } from '../../../client/lib/imperativeModal';
-import UiKitModal from '../../../client/views/modal/uikit/UiKitModal';
-import { sdk } from '../../utils/client/lib/SDKClient';
+import { dispatchToastMessage } from '../../../client/lib/toast';
 import { router } from '../../../client/providers/RouterProvider';
+import { sdk } from '../../utils/client/lib/SDKClient';
+import { t } from '../../utils/lib/i18n';
+
+const UiKitModal = lazy(() => import('../../../client/views/modal/uikit/UiKitModal'));
 
 const events = new Emitter();
 
@@ -45,7 +44,7 @@ export const generateTriggerId = (appId) => {
 	return triggerId;
 };
 
-const handlePayloadUserInteraction = (type, { /* appId,*/ triggerId, ...data }) => {
+export const handlePayloadUserInteraction = (type, { /* appId,*/ triggerId, ...data }) => {
 	if (!triggersId.has(triggerId)) {
 		return;
 	}
@@ -256,11 +255,3 @@ export const getUserInteractionPayloadByViewId = (viewId) => {
 
 	return instance.payload;
 };
-
-Meteor.startup(() =>
-	CachedCollectionManager.onLogin(() =>
-		Notifications.onUser('uiInteraction', ({ type, ...data }) => {
-			handlePayloadUserInteraction(type, data);
-		}),
-	),
-);