+++ apps/meteor/client/views/admin/cloud/components/RegisterWorkspaceMenu.tsx
@@ -1,9 +1,9 @@
-import { Button, ButtonGroup, Icon } from '@rocket.chat/fuselage';
+import { Button, ButtonGroup } from '@rocket.chat/fuselage';
 import { useSetModal, useTranslation } from '@rocket.chat/ui-contexts';
 import React from 'react';
 
 import { useExternalLink } from '../../../../hooks/useExternalLink';
-import { cloudConsoleUrl } from '../constants';
+import { CLOUD_CONSOLE_URL } from '../../../../lib/constants';
 import RegisteredWorkspaceModal from '../modals/RegisteredWorkspaceModal';
 
 type RegisterWorkspaceMenuProps = {
@@ -35,12 +35,10 @@ const RegisterWorkspaceMenu = ({
 		<ButtonGroup>
 			{isWorkspaceRegistered && isConnectedToCloud && (
 				<>
-					<Button role='link' onClick={() => handleLinkClick(cloudConsoleUrl)}>
-						<Icon name='new-window' size='x20' pie={4} />
+					<Button icon='new-window' role='link' onClick={() => handleLinkClick(CLOUD_CONSOLE_URL)}>
 						{t('Cloud')}
 					</Button>
-					<Button onClick={handleManageButton}>
-						<Icon name='customize' size='x20' pie={4} />
+					<Button icon='customize' onClick={handleManageButton}>
 						{t('Manage')}
 					</Button>
 				</>