+++ apps/meteor/client/views/room/Header/Omnichannel/OmnichannelRoomHeader.tsx
@@ -6,8 +6,6 @@ import { useSyncExternalStore } from 'use-sync-external-store/shim';
 
 import BurgerMenu from '../../../../components/BurgerMenu';
 import { useOmnichannelRoom } from '../../contexts/RoomContext';
-import { ToolboxContext, useToolboxContext } from '../../contexts/ToolboxContext';
-import type { ToolboxActionConfig } from '../../lib/Toolbox';
 import RoomHeader from '../RoomHeader';
 import { BackButton } from './BackButton';
 import QuickActions from './QuickActions';
@@ -37,7 +35,6 @@ const OmnichannelRoomHeader: FC<OmnichannelRoomHeaderProps> = ({ slots: parentSl
 
 	const { isMobile } = useLayout();
 	const room = useOmnichannelRoom();
-	const toolbox = useToolboxContext();
 
 	const slots = useMemo(
 		() => ({
@@ -45,27 +42,15 @@ const OmnichannelRoomHeader: FC<OmnichannelRoomHeaderProps> = ({ slots: parentSl
 			start: (!!isMobile || currentRouteName === 'omnichannel-directory' || currentRouteName === 'omnichannel-current-chats') && (
 				<HeaderToolbox>
 					{isMobile && <BurgerMenu />}
-					{<BackButton routeName={currentRouteName} />}
+					<BackButton routeName={currentRouteName} />
 				</HeaderToolbox>
 			),
-			posContent: <QuickActions room={room} />,
+			posContent: <QuickActions />,
 		}),
-		[isMobile, currentRouteName, parentSlot, room],
+		[isMobile, currentRouteName, parentSlot],
 	);
 
-	return (
-		<ToolboxContext.Provider
-			value={useMemo(
-				() => ({
-					...toolbox,
-					actions: new Map([...(Array.from(toolbox.actions.entries()) as [string, ToolboxActionConfig][])]),
-				}),
-				[toolbox],
-			)}
-		>
-			<RoomHeader slots={slots} room={room} />
-		</ToolboxContext.Provider>
-	);
+	return <RoomHeader slots={slots} room={room} />;
 };
 
 export default OmnichannelRoomHeader;