+++ apps/meteor/tests/end-to-end/api/00-autotranslate.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 
 import { getCredentials, api, request, credentials } from '../../data/api-data.js';
-import { updatePermission, updateSetting } from '../../data/permissions.helper';
 import { sendSimpleMessage } from '../../data/chat.helper';
+import { updatePermission, updateSetting } from '../../data/permissions.helper';
 
 describe('AutoTranslate', function () {
 	this.retries(0);