+++ ee/packages/omnichannel-services/src/QueueWorker.ts
@@ -1,8 +1,8 @@
-import type { Db } from 'mongodb';
-import type { Actions, ValidResult, Work } from 'mongo-message-queue';
-import MessageQueue from 'mongo-message-queue';
 import { ServiceClass, api, License } from '@rocket.chat/core-services';
 import type { IQueueWorkerService, HealthAggResult } from '@rocket.chat/core-services';
+import type { Actions, ValidResult, Work } from 'mongo-message-queue';
+import MessageQueue from 'mongo-message-queue';
+import type { Db } from 'mongodb';
 
 import type { Logger } from '../../../../apps/meteor/server/lib/logger/Logger';