+++ packages/core-typings/src/IOmnichannelCannedResponse.ts
@@ -6,7 +6,7 @@ export interface IOmnichannelCannedResponse extends IRocketChatRecord {
 	shortcut: string;
 	text: string;
 	scope: string;
-	tags: any;
+	tags: string[];
 	// userId is optional, its only required when scope === 'user'
 	userId?: IUser['_id'];
 	departmentId?: ILivechatDepartment['_id'];